refactor: Task extracted into its own module
